requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.2
Django==3.2.7
django-decouple==2.1
djangorestframework==3.12.4
filelock==3.0.12
pbr==5.6.0
platformdirs==2.3.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
stevedore==3.4.0
virtualenv==20.8.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
